refactor: 방 타입 수정하기 위한 조회 실패 예외 경로 이동
